+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Adds support for calling `Net.connect` with `port` and `host` arguments.
+
 ## 0.2.0 (Apr 19, 2014)
 - Does not store requests on an instance of `Mitm` any longer.
 - Adds `socket` event to `Mitm`.